Split SaveLoadFunctionality, comment and restructure
