+++ Assets/GracesGames/2DTileMapLevelEditor/Scripts/Functionalities/SaveFunctionality.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 6c67a61cd825343aba742d4225385c49
+timeCreated: 1513780372
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
